style : 공백, 분기문 refactor: Bad Request handler 합침
