+++ src/main/java/com/moviePicker/api/auth/dto/LoginRequest.java
@@ -9,7 +9,6 @@ import lombok.Getter;
 public class LoginRequest {
 
     private final String email;
-
     private final String password;
 
     @JsonCreator
