+++ src/main/java/com/moviePicker/api/auth/dto/LoginResponse.java
@@ -6,8 +6,7 @@ import lombok.Getter;
 
 @Getter
 public class LoginResponse {
-
-
+    
     private final String accessToken;
 
     @JsonCreator
