+++ src/main/java/com/moviePicker/api/member/repository/MemberRepository.java
@@ -6,10 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface MemberRepository extends JpaRepository<Member, String> {
+
     Optional<Member> findByEmail(String email);
-
-
-    boolean existsByEmailOrNickname(String email, String nickName);
 
     boolean existsByEmail(String email);
